chore: rename PlanInfoExtractor
