+++ ksqldb-streams/src/main/java/io/confluent/ksql/execution/streams/KSPlanBuilder.java
@@ -19,11 +19,11 @@ import io.confluent.ksql.GenericKey;
 import io.confluent.ksql.execution.builder.KsqlQueryBuilder;
 import io.confluent.ksql.execution.plan.KGroupedStreamHolder;
 import io.confluent.ksql.execution.plan.KGroupedTableHolder;
+import io.confluent.ksql.execution.plan.PlanInfoExtractor;
 import io.confluent.ksql.execution.plan.PlanInfo;
 import io.confluent.ksql.execution.plan.KStreamHolder;
 import io.confluent.ksql.execution.plan.KTableHolder;
 import io.confluent.ksql.execution.plan.PlanBuilder;
-import io.confluent.ksql.execution.plan.PlanInfoExtractor;
 import io.confluent.ksql.execution.plan.StreamAggregate;
 import io.confluent.ksql.execution.plan.StreamFilter;
 import io.confluent.ksql.execution.plan.StreamFlatMap;
@@ -85,7 +85,7 @@ public final class KSPlanBuilder implements PlanBuilder {
 
   @Override
   public PlanInfoExtractor getPlanInfoExtractor() {
-    return new KSPlanInfoExtractor();
+    return new PlanInfoExtractor();
   }
 
   @Override
